Apply clang-tidy readability-redundant-casting

+++ src/melee/ft/chara/ftCommon/ftCo_CliffJump.c
@@ -33,7 +33,7 @@ bool ftCo_8009B170(Fighter_GObj* gobj)
 void ftCo_8009B1B8(Fighter_GObj* gobj)
 {
     Fighter* fp = gobj->user_data;
-    FtMotionId msid = (float) fp->dmg.x1830_percent < p_ftCommonData->x488
+    FtMotionId msid = fp->dmg.x1830_percent < p_ftCommonData->x488
                           ? ftCo_MS_CliffJumpQuick1
                           : ftCo_MS_CliffJumpSlow1;
     Fighter_ChangeMotionState(gobj, msid, Ft_MF_None, 0, 1, 0, NULL);
