+++ src/melee/ft/chara/ftGameWatch/ftGw_AttackAir.c
@@ -404,7 +404,7 @@ void ftGw_LandingAirN_Enter(HSD_GObj* gobj)
         it_802C73AC(fp->u.gw.x2260_sparkyGObj);
     }
 
-    if ((u32) fp->cmd_vars[0] != 0U) {
+    if (fp->cmd_vars[0] != 0U) {
         ftGw_LandingAirN_Init(gobj);
         fp->accessory4_cb = ftGw_AttackAirN_ItemParachuteOnLand;
         return;
@@ -476,7 +476,7 @@ void ftGw_LandingAirB_Enter(HSD_GObj* gobj)
         it_802C73AC(fp->u.gw.x2260_sparkyGObj);
     }
 
-    if ((u32) fp->cmd_vars[0] != 0) {
+    if (fp->cmd_vars[0] != 0) {
         ftGw_LandingAirB_Init(gobj);
         fp->accessory4_cb = ftGw_AttackAirN_ItemTurtleOnLand;
         return;
